+++ apps/funding-service/src/queue/index.ts
@@ -1,5 +1,5 @@
 import { Signer } from "ethers";
-import { abi } from "@rpch/common";
+import { blockchain } from "@rpch/common";
 import { RequestDB } from "../types";
 import { RequestService } from "../request";
 import { CustomError, createLogger } from "../utils";
@@ -39,12 +39,12 @@ export const checkFreshRequests = async (ops: {
     // check if signer has enough to fund request
     const provider = ops.signer.provider
       ? ops.signer.provider
-      : await abi.getProvider(freshRequest.chain_id);
+      : await blockchain.getProvider(freshRequest.chain_id);
     let connectedSigner = ops.signer.provider
       ? ops.signer
       : ops.signer.connect(provider);
     let address = await connectedSigner.getAddress();
-    const balance = await abi.getBalance(
+    const balance = await blockchain.getBalance(
       constants.SMART_CONTRACTS_PER_CHAIN?.[
         freshRequest.chain_id as keyof typeof constants.SMART_CONTRACTS_PER_CHAIN
       ],
@@ -68,7 +68,7 @@ export const checkFreshRequests = async (ops: {
     });
 
     // sent transaction to fund request
-    const { hash: txHash } = await abi.sendTransaction({
+    const { hash: txHash } = await blockchain.sendTransaction({
       smartContractAddress:
         constants.SMART_CONTRACTS_PER_CHAIN?.[
           freshRequest.chain_id as keyof typeof constants.SMART_CONTRACTS_PER_CHAIN
@@ -88,7 +88,7 @@ export const checkFreshRequests = async (ops: {
       id: freshRequest.id,
     });
     // wait for transaction to reach a certain amount of confirmations
-    const txReceipt = await abi.waitForTransaction(
+    const txReceipt = await blockchain.waitForTransaction(
       txHash,
       provider,
       ops.confirmations
