+++ packages/common/src/index.ts
@@ -3,6 +3,6 @@ export { default as Message } from "./message";
 export { default as Request } from "./request";
 export { default as Response } from "./response";
 export { default as Cache } from "./cache";
-export * as abi from "./abi";
+export * as blockchain from "./blockchain";
 export * as hoprd from "./hoprd";
 export * as utils from "./utils";
